xac-thuc-done-90%

+++ src/main/java/com/nvt/dao/GenericDAO.java
@@ -7,6 +7,6 @@ import com.nvt.mapper.RowMapper;
 public interface GenericDAO<T> {
 	List<T> query(String sql, RowMapper<T> rowMapper, Object... parameters);
 	void update (String sql, Object... parameters);
-	Long insert (String sql, Object... parameters);
+	Integer insert (String sql, Object... parameters);
 	int count(String sql, Object... parameters);
 }
